:recycle: Refactors to fix code quality issues
