+++ opensrp-goldsmith/src/main/java/org/smartregister/goldsmith/configuration/anc/AncFormProcessor.java
@@ -28,7 +28,7 @@ import java.util.List;
 
 import static org.smartregister.chw.anc.util.DBConstants.KEY.BASE_ENTITY_ID;
 import static org.smartregister.family.util.JsonFormUtils.METADATA;
-import static org.smartregister.goldsmith.util.SampleAppJsonFormUtils.populateInjectedFields;
+import static org.smartregister.goldsmith.util.FormProcessorJsonFormUtils.populateInjectedFields;
 import static org.smartregister.util.JsonFormUtils.ENCOUNTER_LOCATION;
 
 public class AncFormProcessor implements ModuleFormProcessor {
