+++ opensrp-goldsmith/src/main/java/org/smartregister/goldsmith/configuration/pnc/PncFormProcessor.java
@@ -36,7 +36,7 @@ import static com.vijay.jsonwizard.constants.JsonFormConstants.Properties.DETAIL
 import static org.smartregister.AllConstants.PLAN_IDENTIFIER;
 import static org.smartregister.chw.anc.util.DBConstants.KEY.BASE_ENTITY_ID;
 import static org.smartregister.family.util.JsonFormUtils.METADATA;
-import static org.smartregister.goldsmith.util.SampleAppJsonFormUtils.populateInjectedFields;
+import static org.smartregister.goldsmith.util.FormProcessorJsonFormUtils.populateInjectedFields;
 import static org.smartregister.util.JsonFormUtils.ENCOUNTER_LOCATION;
 
 public class PncFormProcessor implements ModuleFormProcessor {
